[svn] Find self-extracting archives by their file magic only, not extension/mimetype.
The problem with using extensions/mimetypes for this is that it net way too
many false positives.  Files ending in .com, .bat, etc. would cause the
user to be prompted for recursive extraction.  This makes that problem go
away, and it means that error messages when the user tries to extract a
non-archive .exe will probably be more useful, too.