Update for python 3.8
Thank you for making this software, and making the source available!

This patch has a collection of modifications:

- update to be compatible with python 3.8
- update the tests to include the file list header
- update setup.py (version number bump and add python2.7 + python3.8
classifiers)
- add a dockerfile to preserve the testing environment, using tox as the
test runner
- converted the README to Markdown and added the extension so Github
renders it
- update .gitignore for the tox temporary files

The motivation for this patch-

Ubuntu 20.04 no longer has dtrx in the PPA's, I think because it only
supported python 2. I checked the pypi.org page for dtrx, but I think it
was migrated from pypi.python.org, and the package page is empty.

This patch should be enough to deploy an update to pypi.org, to make the
tool available to anyone via `pip install dtrx`. I've updated the readme
with that information too (and it may be worth adding a badge).

I've deployed a test version of this path to
https://pypi.org/project/dtrx-noahp/