Add a 1-second timeout when running Extraction commands
Instead of endlessly waiting for an extractor to complete, add a timeout
that allows us to interact with the user in the event that the command is
stuck. Right now this does nothing, but we'll be using it to spit out
password prompts.

This also adds a 'timeout_check' to the BaseExtractor class to be overridden
by various extractors. I didn't name it 'check_for_password' because it
could easily be used to do other things.

I tried to maintain python2 compatibility, but this will add a
python-subprocess32 dependency for python2 installations.