Add support for password-protected .7z files
7z files can have passwords. Currently when there's a password, 7z
just hangs indefinitely waiting for input that never comes because
the user never sees the password prompt.

This patch adds support for those password-protected 7z files to
SevenExtractor. It uses a non-blocking read of stdout (NOT stderr;
it's different from rar and zip) and timeouts to see if 7z has sent a
password string, and passes that through to the user.

For the last command in the pipe, we were setting stdout to /dev/null.
Instead, we send stdout output to a pipe and check it in the event of
a timeout.

The strings that I see from p7zip 16.02:

"Enter password (will not be echoed):"
"ERROR: Wrong password : file"  (it only prompts once)
"Everything is Ok"