Add support for password-protected .rar files
Rar files can have passwords. Currently when there's a password, unrar
just fails because stdin is closed when it needs to prompts the user.

This patch adds support for those password-protected rar files. In
order to keep stdin open, we add a new flag in BaseExtractor called
user_stdin that tells run_pipes() whether to override stdin or leave it
alone. This keeps the same behavior for most instances of that class,
but the NoPipeExtractor subclass modifies user_stdin so that derived
classes create objects that leave stdin alone.

We then use the same non-blocking read and timeout infrastructure
that zip files use to check for a password prompt.

The strings that I see from unrar 5.9.4:

"Enter password (will not be echoed) for file:"

"The specified password is incorrect.
Enter password (will not be echoed) for file:"

Since the response is multi-line, we're passing along all (new) output
rather than just the last line.