Don't spit out random text on successful unpacking of passworded archives
Normally stderr contains actual errors, which we display to the user if
they show up. In the case of archives with passwords, this is no longer
the case; stderr is used to prompt the user for a password. So, dtrx
would successfully decrypt an archive, but then spit out the following:

"WARNING:dtrx-log:Error output from this process:

7-Zip [64] 16.02 : Copyright (c) 1999-2016 Igor Pavlov : 2016-05-21
p7zip Version 16.02 (locale=en_US.UTF-8,Utf16=on,HugeFiles=on,64 bits,4 CPUs Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz (406E3),ASM,AES-NI)

Scanning the drive for archives:
1 file, 113272063 bytes (109 MiB)

Extracting archive: /home/dilinger/test.7z

Enter password (will not be echoed):
"

This is obvious confusing and undesirable. Instead, what we do is
change the logging level for stderr when we're dealing with passworded
archives. Only output that stuff w/ the debug logging level (dtrx -vv).

There's one exception - if we hit an ExtractorError exception, that
means there are actual errors waiting in stderr (likely due to an
incorrect password). In that case, we no longer use the debug logging
level for stderr output.