tests: stricter encrypted test + fix
I wanted to grep the output to ensure the error we get is actually from the
lack of password.

Also, for some reason in the test harness, 'stty echo' fails with
"stty: 'standard input': Inappropriate ioctl for device". This leaves
the terminal broken after the encrypted archive failure test, so we
have to manually run stty (specifying stdout) during the cleanup phase.