Switch from deadsnakes to pyenv for base pythons (#38)
Several other changes included:

- update pre-commit hooks
- add vscode repo settings for auto format
- fix an issue with newer `lrzip` tools; it now spits out stderr noise
  unless the `-Q` option is set (previously just `-q` was enough)
- switch from a bash script to invoke for driving the various tests

_Intense_ hacks to support both running locally by mounting the cwd into
the container; need to match host UID+GUID to permit sane read/writing
to the mounted cwd... as well as running in CI, where we need to be root
to clone the repo (why!!!), but we MUST be non-root to pass one of the
tests (`can't write to directory`). Very sad workarounds to get
everything going, I'm not too happy about it but it does work.

Lastly, switch to running ci off a container hosted under dtrx-py
organization on the github container registry.

Invoke task for building + pushing the image. And updated documentation
in the README.